Add textbox widget based on tui-input crate
The exact details will probably change pretty soon. It probably doesn't
make sense to have to hit 'i' to go into insert mode given that it's not
otherwise a modal editor. I think it should either start out in insert
mode or always be active when focused. If it starts in insert mode it
might still be useful to be able to exit and then re-enter, maybe with
enter, but maybe that should be managed externally through the focus.